+++ Makefile
@@ -1,11 +1,11 @@
 test:
-	go test -v
+	go test -v ./...
 
 test-cursor:
-	go test -v -run TestCursor
+	go test -v ./cursor
 
 test-paginator:
-	go test -v -run TestPaginator
+	go test -v ./paginator
 
 test-env-up:
 	docker-compose up -d
@@ -14,5 +14,5 @@ test-env-down:
 	docker-compose down -v
 
 test-coverage:
-	go test --coverprofile=c.out
+	go test --coverprofile=c.out ./...
 	go tool cover -html=c.out
